+++ src/values.ts
@@ -6,6 +6,10 @@ import {
     DeleteValueResponse,
     GetValueParams,
     GetValueResponse,
+    ListContactsParams,
+    ListContactsResponse,
+    ListTransactionsParams,
+    ListTransactionsResponse,
     ListValuesParams,
     ListValuesResponse,
     UpdateValueParams,
@@ -15,8 +19,8 @@ import {formatFilterParams, formatResponse, isSuccessStatus, validateRequiredPar
 import {LightrailRequestError} from "./LightrailRequestError";
 import * as lightrail from "./index";
 import {Value} from "./model";
+import {ContentType} from "./params/ContentType";
 
-// CREATE
 export async function createValue(params: CreateValueParams): Promise<CreateValueResponse> {
     if (!params) {
         throw new Error("params not set");
@@ -26,21 +30,16 @@ export async function createValue(params: CreateValueParams): Promise<CreateValu
 
     const resp = await lightrail.request("POST", "values").send(params);
     if (isSuccessStatus(resp.status)) {
-        return (
-            formatResponse(resp)
-        );
+        return formatResponse(resp);
     }
 
     throw new LightrailRequestError(resp);
 }
 
-// READ
-export async function listValues(params?: ListValuesParams): Promise<ListValuesResponse> {
-    const resp = await lightrail.request("GET", "values").set("accept", (!!params && !!params.csv) ? "text/csv" : "application/json").query(formatFilterParams(params));
+export async function listValues(params?: ListValuesParams, contentType: ContentType = "application/json"): Promise<ListValuesResponse> {
+    const resp = await lightrail.request("GET", "values").set("accept", contentType).query(formatFilterParams(params));
     if (isSuccessStatus(resp.status)) {
-        return (
-            formatResponse(resp)
-        );
+        return formatResponse(resp);
     }
 
     throw new LightrailRequestError(resp);
@@ -51,23 +50,18 @@ export async function getValue(value: string | Value, params?: GetValueParams): 
 
     const resp = await lightrail.request("GET", `values/${encodeURIComponent(valueId)}`).query(params);
     if (isSuccessStatus(resp.status) || resp.status === 404) {
-        return (
-            formatResponse(resp)
-        );
+        return formatResponse(resp);
     }
 
     throw new LightrailRequestError(resp);
 }
 
-// UPDATE
 export async function updateValue(value: string | Value, params: UpdateValueParams): Promise<UpdateValueResponse> {
     const valueId = getValueId(value);
 
     const resp = await lightrail.request("PATCH", `values/${encodeURIComponent(valueId)}`).send(params);
     if (isSuccessStatus(resp.status)) {
-        return (
-            formatResponse(resp)
-        );
+        return formatResponse(resp);
     }
 
     throw new LightrailRequestError(resp);
@@ -78,23 +72,40 @@ export async function changeValuesCode(value: string | Value, params: ChangeValu
 
     const resp = await lightrail.request("POST", `values/${encodeURIComponent(valueId)}/changeCode`).send(params);
     if (isSuccessStatus(resp.status)) {
-        return (
-            formatResponse(resp)
-        );
+        return formatResponse(resp);
     }
 
     throw new LightrailRequestError(resp);
 }
 
-// DELETE
 export async function deleteValue(value: string | Value): Promise<DeleteValueResponse> {
     const valueId = getValueId(value);
 
     const resp = await lightrail.request("DELETE", `values/${encodeURIComponent(valueId)}`);
     if (isSuccessStatus(resp.status) || resp.status === 404) {
-        return (
-            formatResponse(resp)
-        );
+        return formatResponse(resp);
+    }
+
+    throw new LightrailRequestError(resp);
+}
+
+export async function listValuesTransactions(value: string | Value, params?: ListTransactionsParams): Promise<ListTransactionsResponse> {
+    const valueId = getValueId(value);
+
+    const resp = await lightrail.request("GET", `values/${encodeURIComponent(valueId)}/transactions`).query(formatFilterParams(params));
+    if (isSuccessStatus(resp.status)) {
+        return formatResponse(resp);
+    }
+
+    throw new LightrailRequestError(resp);
+}
+
+export async function listValuesAttachedContacts(value: string | Value, params?: ListContactsParams): Promise<ListContactsResponse> {
+    const valueId = getValueId(value);
+
+    const resp = await lightrail.request("GET", `values/${encodeURIComponent(valueId)}/contacts`).query(formatFilterParams(params));
+    if (isSuccessStatus(resp.status)) {
+        return formatResponse(resp);
     }
 
     throw new LightrailRequestError(resp);